pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-support-triage"
version = "0.1.0"
description = "Customer Support Ticket Triage environment for OpenEnv"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.2",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.25.0",
]

[project.scripts]
# Enables: uv run --project . server
server = "support_triage.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["support_triage", "support_triage.server"]

[tool.setuptools.package-dir]
"support_triage" = "."
"support_triage.server" = "server"
